requirements.txt
email-validator
PyJWT
fastapi[all]
PyMySQL
SQLAlchemy
cryptography
python-dotenv
